move controlled gate class definitions into base gate module.
