+++ qiskit/extensions/standard/h.py
@@ -20,8 +20,11 @@ import numpy
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
-from qiskit.qasm import pi
+from qiskit.extensions.standard.s import SGate, SdgGate
+from qiskit.extensions.standard.t import TGate, TdgGate
 from qiskit.extensions.standard.u2 import U2Gate
+from qiskit.extensions.standard.cx import CnotGate
+from qiskit.qasm import pi
 
 
 class HGate(Gate):
@@ -44,6 +47,20 @@ class HGate(Gate):
             definition.append(inst)
         self.definition = definition
 
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return CHGate()
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
+
     def inverse(self):
         """Invert this gate."""
         return HGate()  # self-inverse
@@ -60,3 +77,59 @@ def h(self, q):  # pylint: disable=invalid-name
 
 
 QuantumCircuit.h = h
+
+
+class CHGate(ControlledGate):
+    """controlled-H gate."""
+
+    def __init__(self):
+        """Create new CH gate."""
+        super().__init__("ch", 2, [], num_ctrl_qubits=1)
+        self.base_gate = HGate
+        self.base_gate_name = "h"
+
+    def _define(self):
+        """
+        gate ch a,b {
+            s b;
+            h b;
+            t b;
+            cx a, b;
+            tdg b;
+            h b;
+            sdg b;
+        }
+        """
+        definition = []
+        q = QuantumRegister(2, "q")
+        rule = [
+            (SGate(), [q[1]], []),
+            (HGate(), [q[1]], []),
+            (TGate(), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (TdgGate(), [q[1]], []),
+            (HGate(), [q[1]], []),
+            (SdgGate(), [q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return CHGate()  # self-inverse
+
+    def to_matrix(self):
+        """Return a Numpy.array for the Ch gate."""
+        return np.array([[1, 0, 0, 0],
+                         [0, 1/np.sqrt(2), 0, 1/np.sqrt(2)],
+                         [0, 0, 1, 0],
+                         [0, 1/np.sqrt(2), 0, -1/np.sqrt(2)]], dtype=complex)
+
+
+def ch(self, ctl, tgt):  # pylint: disable=invalid-name
+    """Apply CH from ctl to tgt."""
+    return self.append(CHGate(), [ctl, tgt], [])
+
+
+QuantumCircuit.ch = ch
