+++ qiskit/extensions/standard/rz.py
@@ -16,9 +16,11 @@
 Rotation around the z-axis.
 """
 from qiskit.circuit import Gate
+from qiskit.circuit import ControlledGate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
 from qiskit.extensions.standard.u1 import U1Gate
+from qiskit.extensions.standard.cx import CnotGate
 
 
 class RZGate(Gate):
@@ -41,6 +43,20 @@ class RZGate(Gate):
             definition.append(inst)
         self.definition = definition
 
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return CrzGate(self.params[0])
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
+
     def inverse(self):
         """Invert this gate.
 
@@ -55,3 +71,44 @@ def rz(self, phi, q):  # pylint: disable=invalid-name
 
 
 QuantumCircuit.rz = rz
+
+
+class CrzGate(ControlledGate):
+    """controlled-rz gate."""
+
+    def __init__(self, theta):
+        """Create new crz gate."""
+        super().__init__("crz", 2, [theta], num_ctrl_qubits=1)
+        self.base_gate = RZGate
+        self.base_gate_name = "rz"
+
+    def _define(self):
+        """
+        gate crz(lambda) a,b
+        { u1(lambda/2) b; cx a,b;
+          u1(-lambda/2) b; cx a,b;
+        }
+        """
+        definition = []
+        q = QuantumRegister(2, "q")
+        rule = [
+            (U1Gate(self.params[0] / 2), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (U1Gate(-self.params[0] / 2), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return CrzGate(-self.params[0])
+
+
+def crz(self, theta, ctl, tgt):
+    """Apply crz from ctl to tgt with angle theta."""
+    return self.append(CrzGate(theta), [ctl, tgt], [])
+
+
+QuantumCircuit.crz = crz
