+++ qiskit/extensions/standard/swap.py
@@ -16,10 +16,10 @@
 SWAP gate.
 """
 import numpy
-
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
+from qiskit.extensions.standard.x import ToffoliGate
 from qiskit.extensions.standard.cx import CnotGate
 
 
@@ -45,6 +45,20 @@ class SwapGate(Gate):
             definition.append(inst)
         self.definition = definition
 
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return FredkinGate()
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
+
     def inverse(self):
         """Invert this gate."""
         return SwapGate()  # self-inverse
@@ -63,3 +77,45 @@ def swap(self, qubit1, qubit2):
 
 
 QuantumCircuit.swap = swap
+
+
+class FredkinGate(ControlledGate):
+    """Fredkin gate."""
+
+    def __init__(self):
+        """Create new Fredkin gate."""
+        super().__init__("cswap", 3, [], num_ctrl_qubits=1)
+        self.base_gate = SwapGate
+        self.base_gate_name = "swap"
+
+    def _define(self):
+        """
+        gate cswap a,b,c
+        { cx c,b;
+          ccx a,b,c;
+          cx c,b;
+        }
+        """
+        definition = []
+        q = QuantumRegister(3, "q")
+        rule = [
+            (CnotGate(), [q[2], q[1]], []),
+            (ToffoliGate(), [q[0], q[1], q[2]], []),
+            (CnotGate(), [q[2], q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return FredkinGate()  # self-inverse
+
+
+def cswap(self, ctl, tgt1, tgt2):
+    """Apply Fredkin to circuit."""
+    return self.append(FredkinGate(), [ctl, tgt1, tgt2], [])
+
+
+QuantumCircuit.cswap = cswap
+QuantumCircuit.fredkin = cswap
