+++ qiskit/extensions/standard/u1.py
@@ -19,7 +19,7 @@ import numpy
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit import QuantumRegister
-from qiskit.extensions.standard.u3 import U3Gate
+from qiskit.extensions.standard.cx import CnotGate
 
 
 class U1Gate(Gate):
@@ -39,6 +39,20 @@ class U1Gate(Gate):
             definition.append(inst)
         self.definition = definition
 
+    def control(self, num_ctrl_qubits=1, label=None):
+        """Controlled version of this gate.
+
+        Args:
+            num_ctrl_qubits (int): number of control qubits.
+            label (str or None): An optional label for the gate [Default: None]
+
+        Returns:
+            ControlledGate: controlled version of this gate.
+        """
+        if num_ctrl_qubits == 1:
+            return Cu1Gate(self.params[0])
+        return super().control(num_ctrl_qubits=num_ctrl_qubits, label=label)
+
     def inverse(self):
         """Invert this gate."""
         return U1Gate(-self.params[0])
@@ -56,3 +70,46 @@ def u1(self, theta, q):  # pylint: disable=invalid-name
 
 
 QuantumCircuit.u1 = u1
+
+
+class Cu1Gate(ControlledGate):
+    """controlled-u1 gate."""
+
+    def __init__(self, theta):
+        """Create new cu1 gate."""
+        super().__init__("cu1", 2, [theta], num_ctrl_qubits=1)
+        self.base_gate = U1Gate
+        self.base_gate_name = "u1"
+
+    def _define(self):
+        """
+        gate cu1(lambda) a,b
+        { u1(lambda/2) a; cx a,b;
+          u1(-lambda/2) b; cx a,b;
+          u1(lambda/2) b;
+        }
+        """
+        definition = []
+        q = QuantumRegister(2, "q")
+        rule = [
+            (U1Gate(self.params[0] / 2), [q[0]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (U1Gate(-self.params[0] / 2), [q[1]], []),
+            (CnotGate(), [q[0], q[1]], []),
+            (U1Gate(self.params[0] / 2), [q[1]], [])
+        ]
+        for inst in rule:
+            definition.append(inst)
+        self.definition = definition
+
+    def inverse(self):
+        """Invert this gate."""
+        return Cu1Gate(-self.params[0])
+
+
+def cu1(self, theta, ctl, tgt):
+    """Apply cu1 from ctl to tgt with angle theta."""
+    return self.append(Cu1Gate(theta), [ctl, tgt], [])
+
+
+QuantumCircuit.cu1 = cu1
